// src/app/core/models/parameter/parameter.model.ts
export interface Parameter {
  bicycleRental: boolean;
  solarium: boolean;
  golfCourse: boolean;
  massage: boolean;
  fitnessCentre: boolean;
  freeCancellation: boolean;
  selfCatering: boolean;
  breakfastIncluded: boolean;
  breakfastDinnerIncluded: boolean;
  twinBeds: boolean;
  doubleBed: boolean;
  largeDoubleBed: boolean;
  extraLargeDoubleBed: boolean;
  nonSmokingRooms: boolean;
  parking: boolean;
  veryGoodWifi: boolean;
  familyRooms: boolean;
  swimmingPool: boolean;
  roomService: boolean;
  petsAllowed: boolean;
  facilitesDisabledGuest: boolean;
  spaWellnessCentre: boolean;
  aeroportSchuttle: boolean;
  restaurant: boolean;
  receptionAlwaysOpen: boolean;
  kitchen: boolean;
  privateBathroom: boolean;
  airConditioning: boolean;
  laptopFriendlyWorkspace: boolean;
  bath: boolean;
  privatePool: boolean;
  terrace: boolean;
  balcony: boolean;
  blatScreenTv: boolean;
  washingMachine: boolean;
  patio: boolean;
  soundproofing: boolean;
  viewHotel: boolean;
  seaView: boolean;
  wheelchairAccessible: boolean;
  toiletGrabRails: boolean;
  higherLevelToilet: boolean;
  lowerBathroomSink: boolean;
  emergencyCordBathroom: boolean;
  visualAidsBraille: boolean;
  visualAidsTactileSigns: boolean;
  hotelId: string;
}